feat(filter): Add a filter to check the endpoints of the request
